% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AS_Data.R
\docType{data}
\name{AS_data}
\alias{AS_data}
\title{Data used to generate AS Plot}
\format{
a data.frame with 2 colunms
\describe{
\item{speed}{Instantaneous sprint speed in m/s}
\item{acc}{instantaneous acceleration in m/s^2}
}
}
\source{
Data from a soccer training session using Polar 10hz Catapult Vector 7
}
\usage{
AS_data
}
\description{
Simulation data to illustrate the function.
}
\keyword{datasets}
