% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_data.R
\docType{data}
\name{as_insitu_initial_lm}
\alias{as_insitu_initial_lm}
\title{data generated from prepare_data used to calculate AS Profile}
\format{
a data.frame with 2 colunms
\describe{
\item{speed}{Instantaneous sprint speed in m/s}
\item{acc}{instantaneous acceleration in m/s^2}
}
}
\source{
Data from a soccer training session using Polar 10hz Catapult Vector 7
}
\usage{
as_insitu_initial_lm
}
\description{
Simulation data to illustrate the function.
}
\keyword{datasets}
