% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asprofile-package.R
\docType{package}
\name{asprofile-package}
\alias{asprofile}
\alias{asprofile-package}
\title{asprofile: Acceleration–Speed Profiling for Athlete Monitoring}
\description{
Tools for preparing, modeling, and visualizing in-situ acceleration–speed (A–S) profiles from athlete tracking data. The package implements envelope-based data reduction, confidence-interval and box-filter cleaning methods, and linear A–S model fitting to estimate key performance metrics such as maximal acceleration (A0), maximal sprinting speed (S0), slope, and goodness of fit. Designed for applied sport science research and athlete monitoring workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HL-secondorder/asprofile}
  \item Report bugs at \url{https://github.com/HL-secondorder/asprofile/issues}
}

}
\author{
\strong{Maintainer}: Hanwen Liu \email{liuhanwen102398@gmail.com}

}
\keyword{internal}
