% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{data}
\alias{data}
\title{data simulation data set}
\format{
a data.frame with 2 colunms
\describe{
\item{Speed}{Instantaneous sprint speed in m/s}
\item{acc}{instantaneous acceleration in m/s^2}
}
}
\source{
Data fro a soccer training session using Polar 10hz Team Pro gps
}
\usage{
data
}
\description{
Simulation data to illustrate the function. Use with prepare_data function. This will add a several data frames to the global environment, which is used in the sa_profile function
}
\keyword{datasets}
