% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(file_directory, Velocity, Acceleration)
}
\arguments{
\item{file_directory}{path to folder as character string}

\item{Velocity}{name of the velocity column, by default "Velocity"}

\item{Acceleration}{name of the acceleration column, by default "Acceleration"}
}
\description{
Import csv files from folder and combine in dataframe
}
\details{
This function imports csv files from folder and combine them into a dataframe.
}
