% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSituASProfile.R
\docType{package}
\name{InsituASProfile}
\alias{InSituASProfile}
\alias{InSituASProfile-package}
\title{InsituASProfile: A package for calculating the insitu sprint acceleration-speed profile based on gps data}
\description{
This package calculates the insitu sprint acceleration-speed profile based on gps data using the method described by Morin et al. 2021 published in Journal of Biomechanics, DOI: 10.1016/j.jbiomech.2021.110524
}
\author{
Lasse IshC8i \email{lasse.ishoei@gmail.com}
}
